+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   tags should be moved to the message body.
 - Add SubjectPrefix rule to explicitly check for prefixes in subjects, like
   "fix: bug", "fix!: bug", "fix(scope): bug", and suggest to remove them.
+- Update SubjectCliche to catch more types of clichés, such as only "fix". The
+  check is now also case insensitive, so "Fix", "fix" and "FIX" are all caught.
 
 ## 0.2.0